+++ examples/simple/stream_scene_camera_video.py
@@ -1,12 +1,12 @@
 import cv2
 
-from pupil_labs.realtime_api.simple import discover_one_device
+from pupil_labs.realtime_api.simple import Network
 
 
 def main():
     # Look for devices. Returns as soon as it has found the first device.
     print("Looking for the next best device...")
-    device = discover_one_device(max_search_duration_seconds=10)
+    device = Network().wait_for_new_device(timeout_seconds=10)
     if device is None:
         print("No device found.")
         raise SystemExit(-1)