+++ src/pupil_labs/realtime_api/discovery.py
@@ -1,7 +1,7 @@
 import asyncio
-import functools
 import logging
 import time
+import types
 import typing as T
 
 from zeroconf import ServiceStateChange
@@ -12,6 +12,84 @@ from .models import DiscoveredDeviceInfo
 logger = logging.getLogger(__name__)
 
 
+class Network:
+    def __init__(self) -> None:
+        self._devices = {}
+        self._new_devices = asyncio.Queue()
+        self._aiozeroconf = AsyncZeroconf()
+        self._aiobrowser = AsyncServiceBrowser(
+            self._aiozeroconf.zeroconf,
+            "_http._tcp.local.",
+            handlers=[self._handle_service_change],
+        )
+        self._open = True
+
+    async def close(self) -> None:
+        if self._open:
+            await self._aiobrowser.async_cancel()
+            await self._aiozeroconf.async_close()
+            self._devices.clear()
+            self._devices = None
+            while self._new_devices:
+                self._new_devices.get_nowait()
+            self._aiobrowser = None
+            self._aiozeroconf = None
+            self._open = False
+
+    @property
+    def devices(self) -> T.Tuple[DiscoveredDeviceInfo, ...]:
+        return tuple(self._devices.values())
+
+    async def wait_for_new_device(
+        self, timeout_seconds: T.Optional[float] = None
+    ) -> T.Optional[DiscoveredDeviceInfo]:
+        try:
+            return await asyncio.wait_for(self._new_devices.get(), timeout_seconds)
+        except asyncio.TimeoutError:
+            return None
+
+    def _handle_service_change(
+        self, zeroconf, service_type: str, name: str, state_change: ServiceStateChange
+    ) -> None:
+        logger.debug(f"{state_change} {name}")
+        if is_valid_service_name(name) and state_change in (
+            ServiceStateChange.Added,
+            ServiceStateChange.Updated,
+        ):
+            asyncio.create_task(
+                self._request_info_and_put_new_device(
+                    zeroconf, service_type, name, timeout_ms=3000
+                )
+            )
+        elif name in self._devices:
+            del self._devices[name]
+
+    async def _request_info_and_put_new_device(
+        self, zeroconf, service_type, name, timeout_ms
+    ):
+        info = AsyncServiceInfo(service_type, name)
+        if await info.async_request(zeroconf, timeout_ms):
+            device = DiscoveredDeviceInfo(
+                name,
+                info.server,
+                info.port,
+                ['.'.join([str(symbol) for symbol in addr]) for addr in info.addresses],
+            )
+            self._devices[name] = device
+            await self._new_devices.put(device)
+
+    async def __aenter__(self) -> "Network":
+        return self
+
+    async def __aexit__(
+        self,
+        exc_type: T.Optional[T.Type[BaseException]],
+        exc_val: T.Optional[BaseException],
+        exc_tb: T.Optional[types.TracebackType],
+    ):
+        await self.close()
+
+
 async def discover_devices(
     timeout_seconds: T.Optional[float] = None,
 ) -> T.AsyncIterator[DiscoveredDeviceInfo]:
@@ -20,62 +98,19 @@ async def discover_devices(
     :param timeout_seconds: Stop after ``timeout_seconds``. If ``None``, run discovery
         forever.
     """
-    logger.info("Searching for devices...")
-    async with AsyncZeroconf() as aiozeroconf:
-        queue = asyncio.Queue()
-        queue_fn = functools.partial(_queue_service_state_changes, queue)
-
-        browser = AsyncServiceBrowser(
-            aiozeroconf.zeroconf, "_http._tcp.local.", handlers=[queue_fn]
-        )
-        try:
-            while True:
-                if timeout_seconds is not None and timeout_seconds <= 0.0:
-                    return
-                try:
-                    t0 = time.perf_counter()
-                    yield await asyncio.wait_for(queue.get(), timeout=timeout_seconds)
-                    if timeout_seconds is not None:
-                        timeout_seconds -= time.perf_counter() - t0
-                except asyncio.TimeoutError:
-                    return
-        finally:
-            await browser.async_cancel()
-
-
-async def discover_one_device(
-    max_search_duration_seconds: T.Optional[float] = None,
-) -> T.Optional[DiscoveredDeviceInfo]:
-    """Search until one device is found."""
-    async for dev_info in discover_devices(max_search_duration_seconds):
-        return dev_info
-    return None
+    async with Network() as network:
+        while True:
+            if timeout_seconds is not None and timeout_seconds <= 0.0:
+                return
+            t0 = time.perf_counter()
+            device = await network.wait_for_new_device(timeout_seconds)
+            if device is None:
+                return  # timeout reached
+            else:
+                yield device
+            if timeout_seconds is not None:
+                timeout_seconds -= time.perf_counter() - t0
 
 
 def is_valid_service_name(name: str) -> bool:
     return name.split(":")[0] == "PI monitor"
-
-
-def _queue_service_state_changes(
-    queue, zeroconf, service_type: str, name: str, state_change: ServiceStateChange
-) -> None:
-    if is_valid_service_name(name) and state_change == ServiceStateChange.Added:
-        asyncio.create_task(
-            _request_info_for_valid_services_and_queue_result(
-                queue, zeroconf, service_type, name, timeout_ms=3000
-            )
-        )
-
-
-async def _request_info_for_valid_services_and_queue_result(
-    queue, zeroconf, service_type, name, timeout_ms
-):
-    info = AsyncServiceInfo(service_type, name)
-    await info.async_request(zeroconf, timeout_ms)
-    device = DiscoveredDeviceInfo(
-        name,
-        info.server,
-        info.port,
-        ['.'.join([str(symbol) for symbol in addr]) for addr in info.addresses],
-    )
-    await queue.put(device)